Use TypeScript and Eslint for type aware linting (#119)
* chore: use ts and eslint for type aware linting

* fixup: remove test case for missing git selector in git id

* fixup: revert creation of custom spec fixture for graph tests

* fixup: use prettier ignores for eslint

* fixup: disable no-this-alias rule

* fixup: various lint fixes

* fixup: use unknown as catch type

* fixup: dont require unknown catch in tests

* fixup: coverage

* fixup: auto sort everything and guard consistent rules behind environ

+++ src/rollback-remove/src/index.ts
@@ -15,8 +15,8 @@ export class RollbackRemove {
   async rm(path: string) {
     const target = `${dirname(path)}/.VLT.DELETE.${this.#key}.${basename(path)}`
     this.#paths.set(path, target)
-    await rename(path, target).catch(e => {
-      if (e.code === 'ENOENT') {
+    await rename(path, target).catch((e: unknown) => {
+      if (e instanceof Error && 'code' in e && e.code === 'ENOENT') {
         this.#paths.delete(path)
         return
       }
